decouple data_algebra dependence mark newer versions of incoming packages rebuild and retest

+++ pkg/tests/test_KDD2009.py
@@ -6,13 +6,22 @@ import pytest
 import vtreat
 import vtreat.cross_plan
 import numpy.random
-import data_algebra
-from data_algebra.data_ops import descr, TableDescription
-import vtreat.vtreat_db_adapter
-import data_algebra.BigQuery
 import sklearn.linear_model
 import sklearn.metrics
 import vtreat.stats_utils
+
+have_data_algebra = False
+try:
+    import data_algebra
+    from data_algebra.data_ops import descr, TableDescription
+    import data_algebra.BigQuery
+    import vtreat.vtreat_db_adapter
+    have_data_algebra = True
+except ModuleNotFoundError:
+    pass
+
+
+
 
 
 # From:
@@ -101,40 +110,41 @@ def test_KDD2009_vtreat_1():
     # test transform conversion
     transform_as_data = plan.description_matrix()
     incoming_vars = list(set(transform_as_data['orig_var']))
-    ops = vtreat.vtreat_db_adapter.as_data_algebra_pipeline(
-        source=TableDescription(
-            table_name='d_test',
-            column_names=incoming_vars + ['orig_index']),
-        vtreat_descr=transform_as_data,
-        treatment_table_name='transform_as_data',
-        row_keys=['orig_index'],
-    )
-    test_by_pipeline = ops.eval({
-        'd_test': d_test.loc[:, incoming_vars + ['orig_index']],
-        'transform_as_data': transform_as_data})
-    assert test_by_pipeline.shape[0] == test_processed.shape[0]
-    assert test_by_pipeline.shape[1] >= test_processed.shape[1]
-    assert not numpy.any(numpy.isnan(test_by_pipeline))
-    test_pipeline_cols = set(test_by_pipeline.columns)
-    assert numpy.all([c in test_pipeline_cols for c in test_processed.columns])
-    test_cols_sorted = list(test_processed.columns)
-    test_cols_sorted.sort()
-    assert numpy.abs(test_processed[test_cols_sorted] - test_by_pipeline[test_cols_sorted]).max(axis=0).max() < 1e-5
-    # data algebra pipeline in database
-    sql = data_algebra.BigQuery.BigQueryModel().to_sql(ops)
-    assert isinstance(sql, str)
-    if test_on_BigQuery:
-        db_handle = data_algebra.BigQuery.example_handle()
-        db_handle.drop_table('d_test_processed')
-        db_handle.insert_table(d_test.loc[:, incoming_vars + ['orig_index']], table_name='d_test', allow_overwrite=True)
-        db_handle.insert_table(transform_as_data, table_name='transform_as_data', allow_overwrite=True)
-        db_handle.execute(
-            f"CREATE TABLE {db_handle.db_model.table_prefix}.d_test_processed AS {db_handle.to_sql(ops)}")
-        db_res = db_handle.read_query(
-            f"SELECT * FROM {db_handle.db_model.table_prefix}.d_test_processed ORDER BY orig_index")
-        assert db_res.shape[0] == test_processed.shape[0]
-        assert numpy.abs(test_processed[test_cols_sorted] - db_res[test_cols_sorted]).max(axis=0).max() < 1e-5
-        db_handle.drop_table('d_test')
-        db_handle.drop_table('transform_as_data')
-        db_handle.drop_table('d_test_processed')
-        db_handle.close()
+    if have_data_algebra:
+        ops = vtreat.vtreat_db_adapter.as_data_algebra_pipeline(
+            source=TableDescription(
+                table_name='d_test',
+                column_names=incoming_vars + ['orig_index']),
+            vtreat_descr=transform_as_data,
+            treatment_table_name='transform_as_data',
+            row_keys=['orig_index'],
+        )
+        test_by_pipeline = ops.eval({
+            'd_test': d_test.loc[:, incoming_vars + ['orig_index']],
+            'transform_as_data': transform_as_data})
+        assert test_by_pipeline.shape[0] == test_processed.shape[0]
+        assert test_by_pipeline.shape[1] >= test_processed.shape[1]
+        assert not numpy.any(numpy.isnan(test_by_pipeline))
+        test_pipeline_cols = set(test_by_pipeline.columns)
+        assert numpy.all([c in test_pipeline_cols for c in test_processed.columns])
+        test_cols_sorted = list(test_processed.columns)
+        test_cols_sorted.sort()
+        assert numpy.abs(test_processed[test_cols_sorted] - test_by_pipeline[test_cols_sorted]).max(axis=0).max() < 1e-5
+        # data algebra pipeline in database
+        sql = data_algebra.BigQuery.BigQueryModel().to_sql(ops)
+        assert isinstance(sql, str)
+        if test_on_BigQuery:
+            db_handle = data_algebra.BigQuery.example_handle()
+            db_handle.drop_table('d_test_processed')
+            db_handle.insert_table(d_test.loc[:, incoming_vars + ['orig_index']], table_name='d_test', allow_overwrite=True)
+            db_handle.insert_table(transform_as_data, table_name='transform_as_data', allow_overwrite=True)
+            db_handle.execute(
+                f"CREATE TABLE {db_handle.db_model.table_prefix}.d_test_processed AS {db_handle.to_sql(ops)}")
+            db_res = db_handle.read_query(
+                f"SELECT * FROM {db_handle.db_model.table_prefix}.d_test_processed ORDER BY orig_index")
+            assert db_res.shape[0] == test_processed.shape[0]
+            assert numpy.abs(test_processed[test_cols_sorted] - db_res[test_cols_sorted]).max(axis=0).max() < 1e-5
+            db_handle.drop_table('d_test')
+            db_handle.drop_table('transform_as_data')
+            db_handle.drop_table('d_test_processed')
+            db_handle.close()
